refactoe: duplicate package.json & Jakefile for each modules
In regard to package.json, original is kept under the project root, and duplicated under the root of each
modules.
Inregard to Jakefile, original is not kept but moved under modules/frontend/, and duplicated under
modules/crawler.

In this commit, I just copy package.json with NO modification. The
modification will be done in the latter commit(s).